cli/requirements.txt
httpx>=0.27.0
click>=8.1.0
rich>=13.7.0
anthropic>=0.34.0
openai>=1.40.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
google-generativeai>=0.7.0
mcp>=1.0.0
anyio>=4.4.0
